+++ src/dataframe/dataframe.jl
@@ -98,6 +98,8 @@ use the functionality provided by `select`/`transform`/`combine` functions, use
 functions, or provide type assertions to the variables that hold columns
 extracted from a `DataFrame`.
 
+$MEDATADA_FIXED
+
 # Examples
 ```jldoctest
 julia> DataFrame((a=[1, 2], b=[3, 4])) # Tables.jl table constructor
@@ -165,9 +167,11 @@ julia> DataFrame([1 0; 2 0], :auto) # matrix constructor
    2 │     2      0
 ```
 """
-struct DataFrame <: AbstractDataFrame
+mutable struct DataFrame <: AbstractDataFrame
     columns::Vector{AbstractVector}
     colindex::Index
+    metadata::Union{Nothing, Dict{String, Any}}
+    colmetadata::Union{Nothing, Dict{Int, Dict{String, Any}}}
 
     # the inner constructor should not be used directly
     function DataFrame(columns::Union{Vector{Any}, Vector{AbstractVector}},
@@ -218,7 +222,7 @@ struct DataFrame <: AbstractDataFrame
             firstindex(col) != 1 && _onebased_check_error(i, col)
         end
 
-        new(convert(Vector{AbstractVector}, columns), colindex)
+        new(convert(Vector{AbstractVector}, columns), colindex, nothing, nothing)
     end
 end
 
@@ -284,7 +288,15 @@ function DataFrame(d::AbstractDict; copycols::Bool=true)
     colindex = Index(colnames)
     columns = Any[v for v in values(d)]
     df = DataFrame(columns, colindex, copycols=copycols)
-    d isa Dict && select!(df, sort!(propertynames(df)))
+    if d isa Dict
+        select!(df, sort!(propertynames(df)))
+    else
+        # AbstractDict can potentially implement Tables.jl table interface
+        _copy_metadata!(df, x)
+        for col in _names(df)
+            _copy_colmetadata!(df, col, x, col)
+        end
+    end
     return df
 end
 
@@ -577,13 +589,18 @@ end
     idx = Index(lookup, u)
 
     if length(selected_columns) == 1
-        return DataFrame(AbstractVector[_columns(df)[selected_columns[1]][row_inds]],
-                         idx, copycols=false)
+        new_df = DataFrame(AbstractVector[_columns(df)[selected_columns[1]][row_inds]],
+                           idx, copycols=false)
     else
         # Computing integer indices once for all columns is faster
         selected_rows = T === Bool ? _findall(row_inds) : row_inds
-        _threaded_getindex(selected_rows, selected_columns, _columns(df), idx)
+        new_df = _threaded_getindex(selected_rows, selected_columns, _columns(df), idx)
     end
+    _copy_metadata!(new_df, df)
+    for col in _names(new_df)
+        _copy_colmetadata!(new_df, col, df, col)
+    end
+    return new_df
 end
 
 @inline function Base.getindex(df::DataFrame, row_inds::AbstractVector{T}, ::Colon) where T
@@ -593,12 +610,18 @@ end
     idx = copy(index(df))
 
     if ncol(df) == 1
-        return DataFrame(AbstractVector[_columns(df)[1][row_inds]], idx, copycols=false)
+        new_df = DataFrame(AbstractVector[_columns(df)[1][row_inds]], idx, copycols=false)
     else
         # Computing integer indices once for all columns is faster
         selected_rows = T === Bool ? _findall(row_inds) : row_inds
-        _threaded_getindex(selected_rows, 1:ncol(df), _columns(df), idx)
+        new_df = _threaded_getindex(selected_rows, 1:ncol(df), _columns(df), idx)
     end
+
+    _copy_metadata!(new_df, df)
+    for col in _names(new_df)
+        _copy_colmetadata!(new_df, col, df, col)
+    end
+    return new_df
 end
 
 @inline Base.getindex(df::DataFrame, row_inds::Not, col_inds::MultiColumnIndex) =
@@ -629,6 +652,8 @@ function insert_single_column!(df::DataFrame, v::AbstractVector, col_ind::Column
     if haskey(index(df), col_ind)
         j = index(df)[col_ind]
         _columns(df)[j] = dv
+        # drop metadata if replacing a column
+        _drop_colmetadata!(df, j)
     else
         if col_ind isa SymbolOrString
             push!(index(df), Symbol(col_ind))
@@ -724,6 +749,8 @@ for T1 in (:AbstractVector, :Not, :Colon),
         end
         for (j, col) in enumerate(idxs)
             df[row_inds, col] = new_df[!, j]
+            # copy column metadata from source to target data frame
+            _copy_colmetadata!(df, col, new_df, j)
         end
         return df
     end
@@ -761,6 +788,10 @@ for T1 in (:AbstractVector, :Not, :Colon, :(typeof(!))),
         end
         for (j, col) in enumerate(idxs)
             df[row_inds, col] = (row_inds === !) ? mx[:, j] : view(mx, :, j)
+            # drop column metadata if columns are replaced
+            if row_ind isa typeof(!)
+                _drop_colmetadata!(df, col)
+            end
         end
         return df
     end
@@ -773,9 +804,16 @@ Copy data frame `df`.
 If `copycols=true` (the default), return a new  `DataFrame` holding
 copies of column vectors in `df`.
 If `copycols=false`, return a new `DataFrame` sharing column vectors with `df`.
+
+$MEDATADA_FIXED
 """
 function Base.copy(df::DataFrame; copycols::Bool=true)
-    return DataFrame(copy(_columns(df)), copy(index(df)), copycols=copycols)
+    cdf = DataFrame(copy(_columns(df)), copy(index(df)), copycols=copycols)
+    _copy_metadata!(cdf, df)
+    for col in _names(cdf)
+        _copy_colmetadata!(cdf, col, df, col)
+    end
+    return cdf
 end
 
 """
@@ -786,6 +824,8 @@ Delete rows specified by `inds` from a `DataFrame` `df` in place and return it.
 Internally `deleteat!` is called for all columns so `inds` must be:
 a vector of sorted and unique integers, a boolean vector, an integer,
 or `Not` wrapping any valid selector.
+
+$MEDATADA_FIXED
 
 # Examples
 ```jldoctest
@@ -894,6 +934,8 @@ Internally `deleteat!` is called for all columns so `inds` must be:
 a vector of sorted and unique integers, a boolean vector, an integer,
 or `Not` wrapping any valid selector.
 
+$MEDATADA_FIXED
+
 # Examples
 ```jldoctest
 julia> df = DataFrame(a=1:3, b=4:6)
@@ -950,6 +992,8 @@ keepat!(df::DataFrame, inds::Not) = deleteat!(df, Not(inds))
 
 Remove all rows from `df`, making each of its columns empty.
 
+$MEDATADA_FIXED
+
 # Examples
 ```jldoctest
 julia> df = DataFrame(a=1:3, b=4:6)
@@ -978,6 +1022,8 @@ end
     resize!(df::DataFrame, n::Integer)
 
 Resize `df` to have `n` rows by calling `resize!` on all columns of `df`.
+
+$MEDATADA_FIXED
 
 # Examples
 ```jldoctest
@@ -1017,6 +1063,8 @@ Remove the last row from `df` and return a `NamedTuple` created from this row.
 
     Using this method for very wide data frames may lead to expensive compilation.
 
+$MEDATADA_FIXED
+
 # Examples
 ```jldoctest
 julia> df = DataFrame(a=1:3, b=4:6)
@@ -1052,6 +1100,8 @@ Remove the first row from `df` and return a `NamedTuple` created from this row.
 
     Using this method for very wide data frames may lead to expensive compilation.
 
+$MEDATADA_FIXED
+
 # Examples
 ```jldoctest
 julia> df = DataFrame(a=1:3, b=4:6)
@@ -1085,6 +1135,8 @@ Remove the `i`-th row from `df` and return a `NamedTuple` created from this row.
 !!! note
 
     Using this method for very wide data frames may lead to expensive compilation.
+
+$MEDATADA_FIXED
 
 # Examples
 ```jldoctest
@@ -1128,7 +1180,19 @@ function hcat!(df1::DataFrame, df2::AbstractDataFrame;
     u = add_names(index(df1), index(df2), makeunique=makeunique)
     for i in 1:length(u)
         df1[!, u[i]] = copycols ? df2[:, i] : df2[!, i]
+        _copy_colmetadata!(df1, u[i], df2, i)
     end
+
+    if hasmetadata(df1) === true && hasmetadata(df2) === true
+        meta1 = metadata(df1)
+        meta2 = metadata(df2)
+        for (k, v) in pairs(meta1)
+            if !(haskey(meta2, k) && isequal(meta2, v))
+                delete!(meta1, k)
+            end
+        end
+    end
+
     return df1
 end
 
@@ -1170,6 +1234,8 @@ Convert columns `cols` of data frame `df` from element type `T` to
 `cols` can be any column selector ($COLUMNINDEX_STR; $MULTICOLUMNINDEX_STR).
 
 If `cols` is omitted all columns in the data frame are converted.
+
+$MEDATADA_FIXED
 """
 function allowmissing! end
 
@@ -1211,6 +1277,8 @@ If `cols` is omitted all columns in the data frame are converted.
 
 If `error=false` then columns containing a `missing` value will be skipped instead
 of throwing an error.
+
+$MEDATADA_FIXED
 """
 function disallowmissing! end
 
@@ -1264,6 +1332,8 @@ Update a data frame `df` in-place by repeating its rows. `inner` specifies how m
 times each row is repeated, and `outer` specifies how many times the full set
 of rows is repeated. Columns of `df` are freshly allocated.
 
+$MEDATADA_FIXED
+
 # Example
 ```jldoctest
 julia> df = DataFrame(a=1:2, b=3:4)
@@ -1298,7 +1368,14 @@ julia> df
 function repeat!(df::DataFrame; inner::Integer=1, outer::Integer=1)
     inner < 0 && throw(ArgumentError("inner keyword argument must be non-negative"))
     outer < 0 && throw(ArgumentError("outer keyword argument must be non-negative"))
-    return mapcols!(x -> repeat(x, inner = Int(inner), outer = Int(outer)), df)
+    cols = _columns(df)
+    for (i, col) in enumerate(cols)
+        col_new = repeat(col, inner = Int(inner), outer = Int(outer))
+        firstindex(col_new) != 1 && _onebased_check_error(i, col_new)
+        cols[i] = col_new
+    end
+
+    return df
 end
 
 """
@@ -1306,6 +1383,8 @@ end
 
 Update a data frame `df` in-place by repeating its rows the number of times
 specified by `count`. Columns of `df` are freshly allocated.
+
+$MEDATADA_FIXED
 
 # Example
 ```jldoctest
@@ -1330,7 +1409,13 @@ julia> repeat(df, 2)
 """
 function repeat!(df::DataFrame, count::Integer)
     count < 0 && throw(ArgumentError("count must be non-negative"))
-    return mapcols!(x -> repeat(x, Int(count)), df)
+    cols = _columns(df)
+    for (i, col) in enumerate(cols)
+        col_new = repeat(col, inner = Int(inner), outer = Int(outer))
+        firstindex(col_new) != 1 && _onebased_check_error(i, col_new)
+        cols[i] = col_new
+    end
+    return df
 end
 
 # This is not exactly copy! as in general we allow axes to be different
@@ -1341,6 +1426,16 @@ function _replace_columns!(df::DataFrame, newdf::DataFrame)
     copy!(_columns(df), _columns(newdf))
     copy!(_names(index(df)), _names(newdf))
     copy!(index(df).lookup, index(newdf).lookup)
+
+    meta = getfiled(newdf, :metadata)
+    if meta !== nothing
+        setfield(df, :metadata, copy(meta))
+    end
+    colmeta = getfiled(newdf, :colmetadata)
+    if colmeta !== nothing
+        setfield(df, :colmetadata, copy(colmeta))
+    end
+
     return df
 end
 
@@ -1432,4 +1527,85 @@ function allcombinations(::Type{DataFrame}, pairs::Pair{Symbol, <:Any}...)
     @assert inner == target_rows
     @assert size(out_df) == (target_rows, length(colnames))
     return out_df
+end
+
+"""
+    metadata(df::AbstractDataFrame)
+    metadata(dfr::DataFrameRow)
+    metadata(dfc::DataFrameColumns)
+    metadata(dfr::DataFrameRows)
+    metadata(gdf::GroupedDataFrame)
+
+Return `AbstractDict{String, Any}` storing key-value mappings of table level metadata.
+"""
+function metadata(df::DataFrame)
+    meta = getfield(df, :metadata)
+    meta === nothing || return meta
+    new_meta = Dict{String, Any}()
+    setfield!(df, :metadata, new_meta)
+    return new_meta
+end
+
+"""
+    hasmetadata(df::AbstractDataFrame)
+    hasmetadata(dfr::DataFrameRow)
+    hasmetadata(dfc::DataFrameColumns)
+    hasmetadata(dfr::DataFrameRows)
+    hasmetadata(gdf::GroupedDataFrame)
+
+Return either `Bool`.
+If `false` is returned this means that data frame `df` does not have any table
+level metadata defined. If `true` is returned it means that at table level some
+metadata is defined for `df`.
+"""
+function hasmetadata(df::DataFrame)
+    meta = getfield(df, :metadata)
+    return !(meta === nothing || isempty(meta))
+end
+
+"""
+    colmetadata(df::AbstractDataFrame, col::ColumnIndex)
+    colmetadata(dfr::DataFrameRow, col::ColumnIndex)
+    colmetadata(dfc::DataFrameColumns, col::ColumnIndex)
+    colmetadata(dfr::DataFrameRows, col::ColumnIndex)
+    colmetadata(gdf::GroupedDataFrame, col::ColumnIndex)
+
+Return `AbstractDict{String, Any}` storing
+key-value mappings of column level metadata for column `col`.
+To add or update metadata mutate the returned dictionary.
+
+If `col` is not present in `df` an error is thrown.
+"""
+function colmetadata(df::DataFrame, col::ColumnIndex)
+    idx = index(df)[col]
+    cols_meta = getfield(df, :colmetadata)
+    if cols_meta === nothing
+        meta = Dict{Int, Dict{String, Any}}()
+        setfield!(df, :colmetadata, meta)
+    else
+        meta = cols_meta
+    end
+    return get!(meta, idx) do
+        Dict{String, Any}()
+    end
+end
+
+"""
+    hascolmetadata(df::AbstractDataFrame, col::ColumnIndex)
+    hascolmetadata(dfr::DataFrameRow, col::ColumnIndex)
+    hascolmetadata(dfc::DataFrameColumns, col::ColumnIndex)
+    hascolmetadata(dfr::DataFrameRows, col::ColumnIndex)
+    hascolmetadata(gdf::GroupedDataFrame, col::ColumnIndex)
+
+Return `Bool`.
+If `false` is returned this means that column `col` of data frame `df` does not
+have any column level metadata defined. If `true` is returned it means that for
+`:col` column some metadata is defined.
+
+If `col` is not present in `df` an error is thrown.
+"""
+function hascolmetadata(df::DataFrame, col::ColumnIndex)
+    idx = index(df)[col]
+    meta = getfield(df, :colmetadata)
+    return meta !== nothing && haskey(meta, idx) && !isempty(meta[idx])
 end